# Small Performance improvment for GetBaseKeyFromKeyName method