@jkotas 2020-08-13
Could you please share performance data that demonstrate that this is a measurable performance improvement?

I have doubts that this is improving anything. You are replacing few extra iterations that IndexOf has to do with manual check and extra parameter passing and validation.