@lindexi 2020-08-14
@jkotas Thank you and I added the benchmark code