@lindexi 2020-08-14
> 
> 
> Our users are going to call the public method that are using this private method, not the private method. We prefer to benchmark the public methods that are users are calling. Does this have measurable impact on the public method?

Sorry, I use the private method. Because the public method will read the register and it is slowly.