@stephentoub 2020-09-14
> Because the public method will read the register and it is slowly.

That's @jkotas' point.  Does this change have any meaningful difference on the APIs folks are actually able to invoke?