@stephentoub 2020-10-23
@lindexi, are you planning to do any further work here, or can this be closed?  Thanks for your efforts.