@lindexi 2020-10-24
@stephentoub Thank you and do you think it is necessary to measure the performance of the public  method?