@stephentoub 2020-10-24
Yes, for the same reasons @jkotas outlined in https://github.com/dotnet/runtime/pull/40769#issuecomment-673534240.  There needs to be a measurable improvement a developer would actually see for it to be worth the churn.  Thanks.