@jkotas 2020-10-24
And the change needs to be worth the extra code. We look at two basic performance metrics: code speed and code size.