@danmosemsft 2020-12-14
I'm closing for reasons above. Thanks for the contribution though.